+++ pkg/bloomgateway/config.go
@@ -11,7 +11,7 @@ type Config struct {
 	// Ring configures the ring store used to save and retrieve the different Bloom Gateway instances.
 	// In case it isn't explicitly set, it follows the same behavior of the other rings (ex: using the common configuration
 	// section and the ingester configuration by default).
-	Ring RingConfig `yaml:"ring,omitempty" doc:"description=Defines the ring to be used by the bloom gateway servers and clients. In case this isn't configured, this block supports inheriting configuration from the common ring section."`
+	Ring ring.RingConfig `yaml:"ring,omitempty" doc:"description=Defines the ring to be used by the bloom gateway servers and clients. In case this isn't configured, this block supports inheriting configuration from the common ring section."`
 	// Enabled is the global switch to configures whether Bloom Gateways should be used to filter chunks.
 	Enabled bool `yaml:"enabled"`
 	// Client configures the Bloom Gateway client
@@ -29,7 +29,6 @@ func (cfg *Config) RegisterFlags(f *flag.FlagSet) {
 
 // RegisterFlagsWithPrefix registers flags for the Bloom Gateway configuration with a common prefix.
 func (cfg *Config) RegisterFlagsWithPrefix(prefix string, f *flag.FlagSet) {
-	cfg.Ring.RegisterFlagsWithPrefix(prefix, "collectors/", f)
 	f.BoolVar(&cfg.Enabled, prefix+"enabled", false, "Flag to enable or disable the bloom gateway component globally.")
 	f.IntVar(&cfg.WorkerConcurrency, prefix+"worker-concurrency", 4, "Number of workers to use for filtering chunks concurrently.")
 	f.IntVar(&cfg.MaxOutstandingPerTenant, prefix+"max-outstanding-per-tenant", 1024, "Maximum number of outstanding tasks per tenant.")
@@ -37,17 +36,13 @@ func (cfg *Config) RegisterFlagsWithPrefix(prefix string, f *flag.FlagSet) {
 	// TODO(chaudum): Figure out what the better place is for registering flags
 	// -bloom-gateway.client.* or -bloom-gateway-client.*
 	cfg.Client.RegisterFlags(f)
-}
 
-type RingConfig struct {
-	ring.RingConfigWithRF `yaml:",inline"`
-
-	Tokens int `yaml:"tokens"`
-}
-
-func (cfg *RingConfig) RegisterFlagsWithPrefix(flagsPrefix, storePrefix string, f *flag.FlagSet) {
-	cfg.RingConfigWithRF.RegisterFlagsWithPrefix(flagsPrefix, storePrefix, f)
-	f.IntVar(&cfg.Tokens, flagsPrefix+"ring.tokens", 16, "Number of tokens to use in the ring. The bigger the number of tokens, the more fingerprint ranges the compactor will own, but the smaller these ranges will be. Bigger number of tokens means that more but smaller requests will be handled by each gateway.")
+	// Ring
+	skipFlags := []string{
+		prefix + "ring.tokens",
+	}
+	cfg.Ring.RegisterFlagsWithPrefix(prefix, "collectors/", f, skipFlags...)
+	f.IntVar(&cfg.Ring.NumTokens, prefix+"ring.tokens", 16, "Number of tokens to use in the ring. The bigger the number of tokens, the more fingerprint ranges the compactor will own, but the smaller these ranges will be. Bigger number of tokens means that more but smaller requests will be handled by each gateway.")
 }
 
 type Limits interface {